Add header color and reply reference component
