+++ client/src/components/atoms/InlineCommentReply.vue
@@ -1,31 +1,8 @@
 <template>
   <q-card square class="bg-grey-1 shadow-2 q-mb-md">
-    <q-separator color="blue-1" />
-    <comment-header :comment="comment" />
-    <q-card-section class="q-py-sm">
-      <div v-if="referencedComment" class="q-pl-sm">
-        <small>
-          <q-icon size="sm" name="subdirectory_arrow_right" />
-          <div
-            style="display: inline-block; height: 18px; width: 18px"
-            class="q-mr-sm"
-          >
-            <avatar-image
-              :user="referencedComment.created_by"
-              round
-              class="fit"
-            />
-          </div>
-          <span>
-            <router-link to="#inline-comments">
-              Reply to {{ referencedComment.created_by.username }}
-            </router-link>
-          </span>
-        </small>
-      </div>
-    </q-card-section>
-
-    <q-card-section class="q-py-none">
+    <comment-header :comment="comment" bg-color="#bbe2e8" />
+    <comment-reply-reference :comment="comment" :replies="replies" />
+    <q-card-section>
       <!-- eslint-disable-next-line vue/no-v-html -->
       <div v-html="comment.content" />
     </q-card-section>
@@ -43,11 +20,11 @@
   </q-card>
 </template>
 <script setup>
-import { ref, computed } from "vue"
-import AvatarImage from "./AvatarImage.vue"
+import { ref } from "vue"
+import CommentReplyReference from "./CommentReplyReference.vue"
 import CommentHeader from "./CommentHeader.vue"
 const isReplying = ref(false)
-const props = defineProps({
+defineProps({
   comment: {
     required: true,
     type: Object,
@@ -56,11 +33,5 @@ const props = defineProps({
     required: true,
     type: Array,
   },
-})
-
-const referencedComment = computed(() => {
-  return props.comment.reply_to_id
-    ? props.replies.find((e) => e.id === props.comment.reply_to_id)
-    : null
 })
 </script>
