+++ client/src/components/atoms/OverallCommentReply.vue
@@ -1,31 +1,8 @@
 <template>
   <q-card square class="bg-grey-1 shadow-2 q-mb-md">
-    <q-separator :color="props.isInlineComment ? `blue-1` : `grey-3`" />
-    <comment-header :comment="comment" />
-    <q-card-section class="q-py-sm">
-      <div v-if="referencedComment" class="q-pl-sm">
-        <small>
-          <q-icon size="sm" name="subdirectory_arrow_right" />
-          <div
-            style="display: inline-block; height: 18px; width: 18px"
-            class="q-mr-sm"
-          >
-            <avatar-image
-              :user="referencedComment.created_by"
-              round
-              class="fit"
-            />
-          </div>
-          <span>
-            <router-link to="#overall-comments">
-              Reply to {{ referencedComment.created_by.username }}
-            </router-link>
-          </span>
-        </small>
-      </div>
-    </q-card-section>
-
-    <q-card-section class="q-py-none">
+    <comment-header :comment="comment" bg-color="#eeeeee" />
+    <comment-reply-reference :comment="comment" :replies="replies" />
+    <q-card-section>
       <!--  eslint-disable-next-line vue/no-v-html -->
       <div v-html="comment.content" />
     </q-card-section>
@@ -50,13 +27,13 @@
   </q-card>
 </template>
 <script setup>
-import { computed, ref } from "vue"
-import AvatarImage from "./AvatarImage.vue"
+import { ref } from "vue"
+import CommentReplyReference from "./CommentReplyReference.vue"
 import CommentHeader from "./CommentHeader.vue"
 import CommentEditor from "../forms/CommentEditor.vue"
 const isReplying = ref(false)
 
-const props = defineProps({
+defineProps({
   comment: {
     type: Object,
     required: true,
@@ -73,10 +50,4 @@ function cancelReply() {
 function initiateReply() {
   isReplying.value = true
 }
-
-const referencedComment = computed(() => {
-  return props.comment.reply_to_id
-    ? props.replies.find((e) => e.id === props.comment.reply_to_id)
-    : null
-})
 </script>
